requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
openai
feedparser
beautifulsoup4
newspaper3k
